Add files via upload
applied simple encrypted neural network and federated learning concept on boston housing price problem .
The initial plan was to apply these concepts on financial time series data but federated learning don't work well on the sequence to sequence model.
I would like to thank Udacity , OpenMined,Facebook AI and PySyft for giving this exposure and opportunity.
with the exponential creation of data with time, there would be a humongous amount of data especially on an individual  . this big data would be boon for AI but it will also create an issue of individual privacy, in such situation the privacy-preserving technology came to be very useful. 
I personally think that more work needs to be done in this domain because at present it is full of problem issues on many implementations.